+++ app/Config/Services.php
@@ -29,7 +29,7 @@ class Services extends BaseService
 	//     return new \CodeIgniter\Example();
 	// }
 
-	public static function getSecretKey(){
+	public static function getSecretKey(){		
 		return getenv('JWT_SECRET_KEY');
 	} 
 }