+++ app/Helpers/jwt_helper.php
@@ -6,14 +6,14 @@ use Firebase\JWT\JWT;
 
 function getJWTFromRequest($authenticationHeader): string {
     if(is_null($authenticationHeader)) { //JWT is absent
-        throw new Exceptions\JWTException("Missing or invalid JWT in request");
+        throw new Exception("Missing or invalid JWT in request");
     }
     //JWT is sent from client in the format Bearer XXXXXXXXX
     return explode(' ', $authenticationHeader)[1];
 }
 
 function validateJWTFromRequest(string $encodedToken){
-    $key = Services::getSecreteKey();
+    $key = getenv('JWT_SECRET_KEY');//Services::getSecreteKey();
     $decodedToken = JWT::decode($encodedToken, $key, array('HS256'));
     $userModel = new UserModel();
     $userModel->findUserByEmailAddress($decodedToken->email);
@@ -29,6 +29,6 @@ function getSignedJWTForUser(string $email){
         'exp' => $tokenExpiration,
     ];
 
-    $jwt = JWT::encode($payload, Services::getSecreteKey())
+    $jwt = JWT::encode($payload, Services::getSecreteKey());
     return $jwt;
 }